requirements.txt
Flask==2.3.3
blinker==1.9.0
click==8.1.8
Flask==2.3.3
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
